Custom JS Soundcloud Player Widget created
Yeeeee its ugly but works like a charm. Need's to have a function like
"loadSong(track ID, duration, artworkUrl, waveformUrl)", then our main
client side JS file can use this as a library to interact with the
player. the client side JS file will have to AJAX in the parameters that
it needs to send, then call the function from the player.js library.
also maybe it should look for a <div id="player"></div> tag when the JS
is loaded and populate that tag automatically with the player.